requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
asttokens==2.0.5
backcall==0.2.0
click==8.1.2
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
Flask==2.1.1
Flask-MySQLdb==1.0.1
ipykernel==6.13.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.3.0
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysql==0.0.3
mysqlclient==2.1.0
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pywin32==303
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
